+++ service/cmd/provisionKeycloak.go
@@ -66,7 +66,7 @@ var provisionKeycloakCmd = &cobra.Command{
 	},
 }
 
-var provisionKeycloakFromConfigCmd = &cobra.Command{
+var deprecatedProvisionKeycloakFromConfigCmd = &cobra.Command{
 	Use: "keycloak-from-config",
 	RunE: func(_ *cobra.Command, _ []string) error {
 		slog.Info("Command keycloak-from-config has been deprecated. Please use command 'keycloak' instead.")
@@ -130,5 +130,5 @@ func init() {
 	provisionCmd.AddCommand(provisionKeycloakCmd)
 
 	// Deprecated command
-	provisionCmd.AddCommand(provisionKeycloakFromConfigCmd)
+	provisionCmd.AddCommand(deprecatedProvisionKeycloakFromConfigCmd)
 }