+++ service/entityresolution/keycloak/keycloak_entity_resolution.go
@@ -5,6 +5,7 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"log/slog"
 	"strings"
 
 	"github.com/Nerzal/gocloak/v13"
@@ -19,33 +20,49 @@ import (
 	"google.golang.org/protobuf/types/known/structpb"
 )
 
-const ErrTextCreationFailed = "resource creation failed"
-const ErrTextGetRetrievalFailed = "resource retrieval failed"
-const ErrTextNotFound = "resource not found"
+const (
+	ErrTextCreationFailed     = "resource creation failed"
+	ErrTextGetRetrievalFailed = "resource retrieval failed"
+	ErrTextNotFound           = "resource not found"
+)
 
-const ClientJwtSelector = "azp"
-const UsernameJwtSelector = "preferred_username"
+const (
+	ClientJwtSelector   = "azp"
+	UsernameJwtSelector = "preferred_username"
+)
 
 const serviceAccountUsernamePrefix = "service-account-"
 
 type KeycloakConfig struct {
-	URL            string                 `json:"url"`
-	Realm          string                 `json:"realm"`
-	ClientID       string                 `json:"clientid"`
-	ClientSecret   string                 `json:"clientsecret"`
-	LegacyKeycloak bool                   `json:"legacykeycloak" default:"false"`
-	SubGroups      bool                   `json:"subgroups" default:"false"`
-	InferID        InferredIdentityConfig `json:"inferid,omitempty"`
+	URL            string                 `mapstructure:"url" json:"url"`
+	Realm          string                 `mapstructure:"realm" json:"realm"`
+	ClientID       string                 `mapstructure:"clientid" json:"clientid"`
+	ClientSecret   string                 `mapstructure:"clientsecret" json:"clientsecret"`
+	LegacyKeycloak bool                   `mapstructure:"legacykeycloak" json:"legacykeycloak" default:"false"`
+	SubGroups      bool                   `mapstructure:"subgroups" json:"subgroups" default:"false"`
+	InferID        InferredIdentityConfig `mapstructure:"inferid,omitempty" json:"inferid,omitempty"`
+}
+
+func (c KeycloakConfig) LogValue() slog.Value {
+	return slog.GroupValue(
+		slog.String("url", c.URL),
+		slog.String("realm", c.Realm),
+		slog.String("clientid", c.ClientID),
+		slog.String("clientsecret", "[REDACTED]"),
+		slog.Bool("legacykeycloak", c.LegacyKeycloak),
+		slog.Bool("subgroups", c.SubGroups),
+		slog.Any("inferid", c.InferID),
+	)
 }
 
 type InferredIdentityConfig struct {
-	From EntityImpliedFrom `json:"from,omitempty"`
+	From EntityImpliedFrom `mapstructure:"from,omitempty" json:"from,omitempty"`
 }
 
 type EntityImpliedFrom struct {
-	ClientID bool `json:"clientid,omitempty"`
-	Email    bool `json:"email,omitempty"`
-	Username bool `json:"username,omitempty"`
+	ClientID bool `mapstructure:"clientid,omitempty" json:"clientid,omitempty"`
+	Email    bool `mapstructure:"email,omitempty" json:"email,omitempty"`
+	Username bool `mapstructure:"username,omitempty" json:"username,omitempty"`
 }
 
 type KeyCloakConnector struct {
@@ -59,7 +76,7 @@ func CreateEntityChainFromJwt(
 	kcConfig KeycloakConfig,
 	logger *logger.Logger,
 ) (entityresolution.CreateEntityChainFromJwtResponse, error) {
-	var entityChains = []*authorization.EntityChain{}
+	entityChains := []*authorization.EntityChain{}
 	// for each token in the tokens form an entity chain
 	for _, tok := range req.GetTokens() {
 		entities, err := getEntitiesFromToken(ctx, kcConfig, tok.GetJwt(), logger)
@@ -73,10 +90,8 @@ func CreateEntityChainFromJwt(
 }
 
 func EntityResolution(ctx context.Context,
-	req *entityresolution.ResolveEntitiesRequest, kcConfig KeycloakConfig, logger *logger.Logger) (entityresolution.ResolveEntitiesResponse, error) {
-	// note this only logs when run in test not when running in the OPE engine.
-	logger.Debug("EntityResolution", "req", fmt.Sprintf("%+v", req))
-	// logger.Debug("EntityResolutionConfig", "config", fmt.Sprintf("%+v", kcConfig))
+	req *entityresolution.ResolveEntitiesRequest, kcConfig KeycloakConfig, logger *logger.Logger,
+) (entityresolution.ResolveEntitiesResponse, error) {
 	connector, err := getKCClient(ctx, kcConfig, logger)
 	if err != nil {
 		return entityresolution.ResolveEntitiesResponse{},
@@ -85,22 +100,21 @@ func EntityResolution(ctx context.Context,
 	payload := req.GetEntities()
 
 	var resolvedEntities []*entityresolution.EntityRepresentation
-	logger.Debug("EntityResolution invoked", "payload", payload)
 
 	for idx, ident := range payload {
-		logger.Debug("Lookup", "entity", ident.GetEntityType())
+		logger.Debug("lookup", "entity", ident.GetEntityType())
 		var keycloakEntities []*gocloak.User
 		var getUserParams gocloak.GetUsersParams
 		exactMatch := true
 		switch ident.GetEntityType().(type) {
 		case *authorization.Entity_ClientId:
-			logger.Debug("GetClient", "client_id", ident.GetClientId())
+			logger.Debug("looking up", slog.Any("type", ident.GetEntityType()), slog.String("client_id", ident.GetClientId()))
 			clientID := ident.GetClientId()
 			clients, err := connector.client.GetClients(ctx, connector.token.AccessToken, kcConfig.Realm, gocloak.GetClientsParams{
 				ClientID: &clientID,
 			})
 			if err != nil {
-				logger.Error(err.Error())
+				logger.Error("error getting client info", slog.String("error", err.Error()))
 				return entityresolution.ResolveEntitiesResponse{},
 					status.Error(codes.Internal, ErrTextGetRetrievalFailed)
 			}
@@ -108,13 +122,13 @@ func EntityResolution(ctx context.Context,
 			for _, client := range clients {
 				json, err := typeToGenericJSONMap(client, logger)
 				if err != nil {
-					logger.Error("Error serializing entity representation!", "error", err)
+					logger.Error("error serializing entity representation!", slog.String("error", err.Error()))
 					return entityresolution.ResolveEntitiesResponse{},
 						status.Error(codes.Internal, ErrTextCreationFailed)
 				}
-				var mystruct, structErr = structpb.NewStruct(json)
+				mystruct, structErr := structpb.NewStruct(json)
 				if structErr != nil {
-					logger.Error("Error making struct!", "error", err)
+					logger.Error("error making struct!", slog.String("error", structErr.Error()))
 					return entityresolution.ResolveEntitiesResponse{},
 						status.Error(codes.Internal, ErrTextCreationFailed)
 				}
@@ -124,7 +138,7 @@ func EntityResolution(ctx context.Context,
 				// convert entity to json
 				entityStruct, err := entityToStructPb(ident)
 				if err != nil {
-					logger.Error("unable to make entity struct", "error", err)
+					logger.Error("unable to make entity struct", slog.String("error", err.Error()))
 					return entityresolution.ResolveEntitiesResponse{}, status.Error(codes.Internal, ErrTextCreationFailed)
 				}
 				jsonEntities = append(jsonEntities, entityStruct)
@@ -157,12 +171,12 @@ func EntityResolution(ctx context.Context,
 				status.Error(codes.Internal, ErrTextGetRetrievalFailed)
 		case len(users) == 1:
 			user := users[0]
-			logger.Debug("User found", "user", *user.ID, "entity", ident.String())
-			logger.Debug("User", "details", fmt.Sprintf("%+v", user))
-			logger.Debug("User", "attributes", fmt.Sprintf("%+v", user.Attributes))
+			logger.Debug("user found", slog.String("user", *user.ID), slog.String("entity", ident.String()))
+			logger.Debug("user", slog.Any("details", user))
+			logger.Debug("user", slog.Any("attributes", user.Attributes))
 			keycloakEntities = append(keycloakEntities, user)
 		default:
-			logger.Error("No user found for", "entity", ident)
+			logger.Error("no user found for", slog.Any("entity", ident))
 			if ident.GetEmailAddress() != "" { //nolint:nestif // this case has many possible outcomes to handle
 				// try by group
 				groups, groupErr := connector.client.GetGroups(
@@ -173,11 +187,11 @@ func EntityResolution(ctx context.Context,
 				)
 				switch {
 				case groupErr != nil:
-					logger.Error("Error getting group", "group", groupErr)
+					logger.Error("error getting group", slog.String("group", groupErr.Error()))
 					return entityresolution.ResolveEntitiesResponse{},
 						status.Error(codes.Internal, ErrTextGetRetrievalFailed)
 				case len(groups) == 1:
-					logger.Info("Group found for", "entity", ident.String())
+					logger.Info("group found for", slog.String("entity", ident.String()))
 					group := groups[0]
 					expandedRepresentations, exErr := expandGroup(ctx, *group.ID, connector, &kcConfig, logger)
 					if exErr != nil {
@@ -187,7 +201,7 @@ func EntityResolution(ctx context.Context,
 						keycloakEntities = expandedRepresentations
 					}
 				default:
-					logger.Error("No group found for", "entity", ident.String())
+					logger.Error("no group found for", slog.String("entity", ident.String()))
 					var entityNotFoundErr entityresolution.EntityNotFoundError
 					switch ident.GetEntityType().(type) {
 					case *authorization.Entity_EmailAddress:
@@ -198,7 +212,7 @@ func EntityResolution(ctx context.Context,
 					// 	return &entityresolution.IdpPluginResponse{},
 					// 		status.Error(codes.InvalidArgument, db.ErrTextNotFound)
 					default:
-						logger.Error("Unsupported/unknown type for", "entity", ident.String())
+						logger.Error("unsupported/unknown type for", slog.String("entity", ident.String()))
 						entityNotFoundErr = entityresolution.EntityNotFoundError{Code: int32(codes.NotFound), Message: ErrTextGetRetrievalFailed, Entity: ident.String()}
 					}
 					logger.Error(entityNotFoundErr.String())
@@ -206,7 +220,7 @@ func EntityResolution(ctx context.Context,
 						// user not found -- add json entity to resp instead
 						entityStruct, err := entityToStructPb(ident)
 						if err != nil {
-							logger.Error("unable to make entity struct from email or username", "error", err)
+							logger.Error("unable to make entity struct from email or username", slog.String("error", err.Error()))
 							return entityresolution.ResolveEntitiesResponse{}, status.Error(codes.Internal, ErrTextCreationFailed)
 						}
 						jsonEntities = append(jsonEntities, entityStruct)
@@ -219,7 +233,7 @@ func EntityResolution(ctx context.Context,
 					// user not found -- add json entity to resp instead
 					entityStruct, err := entityToStructPb(ident)
 					if err != nil {
-						logger.Error("unable to make entity struct from username", "error", err)
+						logger.Error("unable to make entity struct from username", slog.String("error", err.Error()))
 						return entityresolution.ResolveEntitiesResponse{}, status.Error(codes.Internal, ErrTextCreationFailed)
 					}
 					jsonEntities = append(jsonEntities, entityStruct)
@@ -233,13 +247,13 @@ func EntityResolution(ctx context.Context,
 		for _, er := range keycloakEntities {
 			json, err := typeToGenericJSONMap(er, logger)
 			if err != nil {
-				logger.Error("Error serializing entity representation!", "error", err)
+				logger.Error("error serializing entity representation!", slog.String("error", err.Error()))
 				return entityresolution.ResolveEntitiesResponse{},
 					status.Error(codes.Internal, ErrTextCreationFailed)
 			}
-			var mystruct, structErr = structpb.NewStruct(json)
+			mystruct, structErr := structpb.NewStruct(json)
 			if structErr != nil {
-				logger.Error("Error making struct!", "error", err)
+				logger.Error("error making struct!", slog.String("error", structErr.Error()))
 				return entityresolution.ResolveEntitiesResponse{},
 					status.Error(codes.Internal, ErrTextCreationFailed)
 			}
@@ -254,9 +268,10 @@ func EntityResolution(ctx context.Context,
 			resolvedEntities,
 			&entityresolution.EntityRepresentation{
 				OriginalId:      originialID,
-				AdditionalProps: jsonEntities},
+				AdditionalProps: jsonEntities,
+			},
 		)
-		logger.Debug("Entities", "resolved", fmt.Sprintf("%+v", resolvedEntities))
+		logger.Debug("entities", slog.Any("resolved", resolvedEntities))
 	}
 
 	return entityresolution.ResolveEntitiesResponse{
@@ -268,14 +283,14 @@ func typeToGenericJSONMap[Marshalable any](inputStruct Marshalable, logger *logg
 	// For now, since we dont' know the "shape" of the entity/user record or representation we will get from a specific entity store,
 	tmpDoc, err := json.Marshal(inputStruct)
 	if err != nil {
-		logger.Error("Error marshalling input type!", "error", err)
+		logger.Error("error marshalling input type!", slog.String("error", err.Error()))
 		return nil, err
 	}
 
 	var genericMap map[string]interface{}
 	err = json.Unmarshal(tmpDoc, &genericMap)
 	if err != nil {
-		logger.Error("Could not deserialize generic entitlement context JSON input document!", "error", err)
+		logger.Error("could not deserialize generic entitlement context JSON input document!", slog.String("error", err.Error()))
 		return nil, err
 	}
 
@@ -285,7 +300,7 @@ func typeToGenericJSONMap[Marshalable any](inputStruct Marshalable, logger *logg
 func getKCClient(ctx context.Context, kcConfig KeycloakConfig, logger *logger.Logger) (*KeyCloakConnector, error) {
 	var client *gocloak.GoCloak
 	if kcConfig.LegacyKeycloak {
-		logger.Warn("Using legacy connection mode for Keycloak < 17.x.x")
+		logger.Warn("using legacy connection mode for Keycloak < 17.x.x")
 		client = gocloak.NewClient(kcConfig.URL)
 	} else {
 		client = gocloak.NewClient(kcConfig.URL, gocloak.SetAuthAdminRealms("admin/realms"), gocloak.SetAuthRealms("realms"))
@@ -303,7 +318,7 @@ func getKCClient(ctx context.Context, kcConfig KeycloakConfig, logger *logger.Lo
 	// logger.Debug(kcConfig.Realm)
 	token, err := client.LoginClient(ctx, kcConfig.ClientID, kcConfig.ClientSecret, kcConfig.Realm)
 	if err != nil {
-		logger.Error("Error connecting to keycloak!", "error", err)
+		logger.Error("error connecting to keycloak!", slog.String("error", err.Error()))
 		return nil, err
 	}
 	keycloakConnector := KeyCloakConnector{token: token, client: client}
@@ -312,7 +327,7 @@ func getKCClient(ctx context.Context, kcConfig KeycloakConfig, logger *logger.Lo
 }
 
 func expandGroup(ctx context.Context, groupID string, kcConnector *KeyCloakConnector, kcConfig *KeycloakConfig, logger *logger.Logger) ([]*gocloak.User, error) {
-	logger.Info("expandGroup invoked: ", groupID, kcConnector, kcConfig.URL, ctx)
+	logger.Info("expanding group", slog.String("groupID", groupID))
 	var entityRepresentations []*gocloak.User
 
 	grp, err := kcConnector.client.GetGroup(ctx, kcConnector.token.AccessToken, kcConfig.Realm, groupID)
@@ -320,16 +335,16 @@ func expandGroup(ctx context.Context, groupID string, kcConnector *KeyCloakConne
 		grpMembers, memberErr := kcConnector.client.GetGroupMembers(ctx, kcConnector.token.AccessToken, kcConfig.Realm,
 			*grp.ID, gocloak.GetGroupsParams{})
 		if memberErr == nil {
-			logger.Debug("Adding members", "amount", len(grpMembers), "from group", *grp.Name)
+			logger.Debug("adding members", slog.Int("amount", len(grpMembers)), slog.String("from group", *grp.Name))
 			for i := 0; i < len(grpMembers); i++ {
 				user := grpMembers[i]
 				entityRepresentations = append(entityRepresentations, user)
 			}
 		} else {
-			logger.Error("Error getting group members", "error", memberErr)
+			logger.Error("error getting group members", slog.String("error", memberErr.Error()))
 		}
 	} else {
-		logger.Error("Error getting group", "error", err)
+		logger.Error("error getting group", slog.String("error", err.Error()))
 		return nil, err
 	}
 	return entityRepresentations, nil
@@ -344,8 +359,8 @@ func getEntitiesFromToken(ctx context.Context, kcConfig KeycloakConfig, jwtStrin
 	if err != nil {
 		return nil, errors.New("error getting claims from jwt")
 	}
-	var entities = []*authorization.Entity{}
-	var entityID = 0
+	entities := []*authorization.Entity{}
+	entityID := 0
 
 	// extract azp
 	extractedValue, okExtract := claims[ClientJwtSelector]
@@ -359,7 +374,8 @@ func getEntitiesFromToken(ctx context.Context, kcConfig KeycloakConfig, jwtStrin
 	entities = append(entities, &authorization.Entity{
 		EntityType: &authorization.Entity_ClientId{ClientId: extractedValueCasted},
 		Id:         fmt.Sprintf("jwtentity-%d", entityID),
-		Category:   authorization.Entity_CATEGORY_ENVIRONMENT})
+		Category:   authorization.Entity_CATEGORY_ENVIRONMENT,
+	})
 	entityID++
 
 	extractedValueUsername, okExp := claims[UsernameJwtSelector]
@@ -381,19 +397,22 @@ func getEntitiesFromToken(ctx context.Context, kcConfig KeycloakConfig, jwtStrin
 			entities = append(entities, &authorization.Entity{
 				EntityType: &authorization.Entity_ClientId{ClientId: clientid},
 				Id:         fmt.Sprintf("jwtentity-%d", entityID),
-				Category:   authorization.Entity_CATEGORY_SUBJECT})
+				Category:   authorization.Entity_CATEGORY_SUBJECT,
+			})
 		} else {
 			// if the returned clientId is empty, no client found, its not a serive account proceed with username
 			entities = append(entities, &authorization.Entity{
 				EntityType: &authorization.Entity_UserName{UserName: extractedValueUsernameCasted},
 				Id:         fmt.Sprintf("jwtentity-%d", entityID),
-				Category:   authorization.Entity_CATEGORY_SUBJECT})
+				Category:   authorization.Entity_CATEGORY_SUBJECT,
+			})
 		}
 	} else {
 		entities = append(entities, &authorization.Entity{
 			EntityType: &authorization.Entity_UserName{UserName: extractedValueUsernameCasted},
 			Id:         fmt.Sprintf("jwtentity-%d", entityID),
-			Category:   authorization.Entity_CATEGORY_SUBJECT})
+			Category:   authorization.Entity_CATEGORY_SUBJECT,
+		})
 	}
 
 	return entities, nil
@@ -415,12 +434,12 @@ func getServiceAccountClient(ctx context.Context, username string, kcConfig Keyc
 		return "", err
 	case len(clients) == 1:
 		client := clients[0]
-		logger.Debug("Client found", "client", *client.ClientID)
+		logger.Debug("client found", slog.String("client", *client.ClientID))
 		return *client.ClientID, nil
 	case len(clients) > 1:
-		logger.Error("More than one client found for ", "clientid", expectedClientName)
+		logger.Error("more than one client found for ", slog.String("clientid", expectedClientName))
 	default:
-		logger.Debug("No client found, likely not a service account", "clientid", expectedClientName)
+		logger.Debug("no client found, likely not a service account", slog.String("clientid", expectedClientName))
 	}
 
 	return "", nil